[v5] [icons] Defer loading of webpack loaders
